requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
earthengine-api==0.1.385
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
numpy==1.24.3
tensorflow==2.15.0
Pillow==10.1.0
